requirements.txt
Django>=2.1.3
requests>=2.20.1
psycopg2>=2.7.6.1
djangorestframework>=3.9.0
